changed some file names to be correspondingwith the whole code
