--- src/models/user.model.ts
@@ -1,34 +1,0 @@
-import mongoose from "mongoose";
-import { User } from "../types/user.type";
-
-const userSchema = new mongoose.Schema<User>({
-  email: {
-    type: String,
-    required: true,
-  },
-  username: {
-    type: String,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  firstName: {
-    type: String,
-    required: true,
-  },
-  lastName: {
-    type: String,
-    required: true,
-  },
-  imgUrl: {
-    type: String,
-  },
-  refreshTokens: {
-    type: [String],
-    required: false,
-  },
-});
-
-export default mongoose.model<User>("User", userSchema);
